Fix commit author collector for Git based repositories
GIT allows us to filter the commit history by author. Since the purpose
of this tool is to render a calendar showing the contributions per user
it makes sense to include the filter to separate the commits created by
the user (registered at git/config) from the rest of the other authors.